tcp、udp、进程、线程跨平台优化
